+++ gpaw/response/integrators.py
@@ -14,7 +14,7 @@ from gpaw.utilities.blas import gemm, rk, czher, mmm
 from gpaw.utilities.progressbar import ProgressBar
 from functools import partial
 
- 
+
 class Integrator():
 
     def __init__(self, cell_cv, comm=mpi.world,
@@ -45,8 +45,6 @@ class Integrator():
 
         if comm.rank != 0:
             txt = devnull
-        elif isinstance(txt, str):
-            txt = open(txt, 'w')
         self.fd = convert_string_to_fd(txt, comm)
 
         self.timer = timer or Timer()
@@ -240,7 +238,7 @@ class PointIntegrator(Integrator):
             else:
                 x_m = 2 * deps_m / (omega.imag**2 + deps_m**2)
                 mynx_mG = n_mG[:, self.Ga:self.Gb] * x_m[:, np.newaxis]
-                mmm(1.0, mynx_mG, 'c', n_mG, 'n', 1.0, chi0_wGG[w])
+                mmm(1.0, mynx_mG, 'C', n_mG, 'N', 1.0, chi0_wGG[w])
 
     @timer('CHI_0 spectral function update')
     def update_hilbert(self, n_mG, deps_m, wd, chi0_wGG):
@@ -294,7 +292,7 @@ class PointIntegrator(Integrator):
         else:
             deps1_m = deps_m + 1j * eta
             deps2_m = deps_m - 1j * eta
-                    
+
         for w, omega in enumerate(omega_w):
             x_m = (1 / (omega + deps1_m) - 1 / (omega - deps2_m))
             chi0_wxvG[w, 0] += np.dot(x_m * n_mG[:, :3].T, n_mG.conj())
@@ -312,7 +310,7 @@ class PointIntegrator(Integrator):
                 break
             o1, o2 = omega_w[w:w + 2]
             assert o1 <= o <= o2, (o1, o, o2)
-            
+
             p = 1 / (o2 - o1)**2
             p1 = p * (o2 - o)
             p2 = p * (o - o1)