+++ gpaw/spline.py
@@ -24,6 +24,7 @@ class Spline:
         # Copy so we don't change the values of the input array
         f_g[-1] = 0.0
         self.spline = _gpaw.Spline(l, rmax, f_g)
+        self.l = self.get_angular_momentum_number()
 
     def get_cutoff(self):
         """Return the radial cutoff."""
@@ -75,6 +76,6 @@ class Spline:
 ##         self.rcut = r_g[-1]
 ##         self.l = l
 ##         ...
-        
+
 ##     def __call__(self, r):
 ##         return self.spline(r)*r**l