+++ gpaw/tddft/__init__.py
@@ -151,7 +151,7 @@ class TDDFT(GPAW):
 
         # Time-dependent variables and operators
         self.td_potential = td_potential
-        self.td_hamiltonian = TimeDependentHamiltonian(self.wfs, self.atoms,
+        self.td_hamiltonian = TimeDependentHamiltonian(self.wfs, self.spos_ac,
                                                        self.hamiltonian,
                                                        td_potential)
         self.td_overlap = self.wfs.overlap  # TODO remove this property
@@ -517,7 +517,7 @@ class TDDFT(GPAW):
         self.kick_strength = np.array(kick_strength)
 
         abs_kick_hamiltonian = AbsorptionKickHamiltonian(
-            self.wfs, self.atoms,
+            self.wfs, self.spos_ac,
             np.array(kick_strength, float))
         abs_kick = AbsorptionKick(self.wfs, abs_kick_hamiltonian,
                                   self.td_overlap, self.solver,