+++ gpaw/test/poisson/poisson.py
@@ -14,7 +14,6 @@ def f(n, p):
     print(a.shape)
     #p = PoissonSolver(nn=1, relax=relax)
     p.set_grid_descriptor(gd)
-    p.initialize()
     cut = N / 2.0 * 0.9
     s = Spline(l=0, rmax=cut, f_g=np.array([1, 0.5, 0.0]))
     c = LFC(gd, [[s], [s]])