+++ gpaw/xc/functional.py
@@ -57,7 +57,6 @@ class XCFunctional(object):
 
     def calculate_paw_correction(self, setup, D_sp, dEdD_sp=None, a=None):
         raise NotImplementedError
-        #return setup.xc_correction.calculate(self, D_sp, dEdD_sp)
 
     def set_positions(self, spos_ac, atom_partition=None):
         pass